docshell/test/mochitest/test_bug511449.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=511449
-->
<head>
  <title>Test for Bug 511449</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="/tests/SimpleTest/NativeKeyCodes.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=511449">Mozilla Bug 511449</a>
<p id="display"></p>
<div id="status"></div>
<div id="content">
</div>
<input type="text" id="input">
<pre id="test">
<script type="application/javascript">

/** Test for Bug 511449 **/

SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("untriaged");
window.addEventListener("load", runTest);

var win = null;

function runTest() {
  document.getElementById("input").focus();
  win = window.open("file_bug511449.html", "");
  SimpleTest.waitForFocus(runNextTest, win);
}

function runNextTest() {
  var didClose = false;
  win.onunload = function() {
    didClose = true;
  };
  synthesizeNativeKey(KEYBOARD_LAYOUT_EN_US, MAC_VK_ANSI_W, {metaKey: 1}, "w", "w");

  setTimeout(function() {
    ok(didClose, "Cmd+W should have closed the tab");
    if (!didClose) {
      win.close();
    }
    SimpleTest.finish();
  }, 1000);
}

</script>

</body>
</html>
